% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootin.R
\name{boot.infit}
\alias{boot.infit}
\title{Bootstrap boot infit.mnsq Function}
\usage{
boot.infit(object)
}
\arguments{
\item{object:}{the name of objectset}
}
\value{
: Bootstrap Statistics
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{

   #boot.infit(object)

}
