% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfit.R
\name{boot.out}
\alias{boot.out}
\title{Bootstrap outfit.mnsq Function}
\usage{
boot.out(object, indices)
}
\arguments{
\item{object:}{the name of objectset}
}
\value{
: Bootstrap Item Outfit.mnsq fit statistis
}
\description{
This function calculates bootstrap item Outfit Mnsq using MML estimation
}
\examples{
   #Not run
   #object <- read.csv("object.csv")
   #outfit.mnsq(object)

}
