% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootout.R
\name{boot.outfit}
\alias{boot.outfit}
\title{Bootstrap boot outfit.mnsq Function}
\usage{
boot.outfit(object)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq fit statistis
}
\description{
This function calculates bootstrap item outfit Mnsq using MML estimation
}
\examples{
   #Not run
   #boot.outfit(data)

}
