% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infitci.R
\name{infit.con}
\alias{infit.con}
\title{Bootstrap boot infit.mnsq CI Function}
\usage{
infit.con(data)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #boot.infit(data)

}
