% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infitConfi.R
\name{infit.conf}
\alias{infit.conf}
\title{Bootstrap boot infit.mnsq CI Function}
\usage{
infit.conf(object)
}
\arguments{
\item{data:}{Data Frame}
}
\value{
: boot infit.mnsq CI Function
}
\description{
This function calculates bootstrap item infit 95% Confidence Interval
using 'boot' package
}
\examples{
   #Not run
   #infit.conf(data)

}
