% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infit.R, R/infitci.R
\name{infit.mnsq}
\alias{infit.mnsq}
\title{Bootstrap Infit.mnsq Function}
\usage{
infit.mnsq(data, indices)

infit.mnsq(data, indices)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Infit.mnsq fit statistis

: print Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item Infit Mnsq using MML estimation

This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
    #Not run
    #data <- read.csv("data.csv")
   #infit.mnsq(data)

   #Not run
   #infit.conf(data)

}
