% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfitConfi.R
\name{outfit.conf}
\alias{outfit.conf}
\title{Bootstrap outfit.mnsq CI Function}
\usage{
outfit.conf(object)
}
\arguments{
\item{data:}{Data Frame}
}
\value{
: bootstrap 95% Confidence Interval  of outfit statistics
}
\description{
This function calculates bootstrap 95% Confidence Interval
of outfit statistics using 'boot'package
}
\examples{
   #Not run
   #outfit.conf(object)

}
