% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infit.R
\name{boot.in}
\alias{boot.in}
\title{Bootstrap Infit.mnsq Function}
\usage{
boot.in(object, indices)
}
\arguments{
\item{object:}{Data Frame}
}
\value{
: Bootstrap Item Infit.mnsq fit statistis
}
\description{
This function calculates bootstrap item Infit Mnsq using MML estimation
}
\examples{
    #Not run
    #object <- read.csv("object.csv")
   #infit.mnsq(object)

}
