% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootin.R
\name{boot.infit}
\alias{boot.infit}
\title{Bootstrap boot infit.mnsq Function}
\usage{
boot.infit(object)
}
\arguments{
\item{object:}{Data frame}
}
\value{
: Bootstrap t Statistics. Also see 'boot' package
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
for the Rasch unidimensional model.
}
\examples{

   #boot.infit(object)

}
