% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootin.R
\name{boot.infit}
\alias{boot.infit}
\title{Bootstrap boot infit.mnsq Function}
\usage{
boot.infit(data)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq fit statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #boot.infit(data)

}
