% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infitConfi.R
\name{infit.conf}
\alias{infit.conf}
\title{Bootstrap boot infit.mnsq CI Function}
\usage{
infit.conf(object)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #infit.conf(data)

}
