% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infitci.R
\name{infit.conf}
\alias{infit.conf}
\title{Bootstrap boot infit.mnsq CI Function}
\usage{
infit.conf(object)
}
\arguments{
\item{object:}{the name of objectset}
}
\value{
: print Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #infit.conf(object)

}
