% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infitci.R
\name{infit.confi}
\alias{infit.confi}
\title{Bootstrap boot infit.mnsq CI Function}
\usage{
infit.confi(data)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: print Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #infit.conf(data)

}
