% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfitci.R
\name{outfit.conf}
\alias{outfit.conf}
\title{Bootstrap boot outfit.mnsq CI Function}
\usage{
outfit.conf(data)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq CI statistis
}
\description{
This function calculates bootstrap item infit Mnsq using MML estimation
}
\examples{
   #Not run
   #outfit.conf(data)

}
