% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfit.R
\name{outfit.mnsq}
\alias{outfit.mnsq}
\title{Bootstrap outfit.mnsq Function}
\usage{
outfit.mnsq(data, indices)
}
\arguments{
\item{data:}{the name of dataset}
}
\value{
: Bootstrap Item Outfit.mnsq fit statistis
}
\description{
This function calculates bootstrap item Outfit Mnsq using MML estimation
}
\examples{
   #Not run
   #data <- read.csv("data.csv")
   #outfit.mnsq(data)

}
